Refactored Logic
